New TaintLevel logic for tag propagation
For testing purposes, added the option to provide a *<TaintLevel>
parameter to each taint source in the sources file. When each source is
read in, if no * is found, the source is assumed to have a level of
TAINT. However, additional levels may be specified after the *, e.g.,
*MAYBE_TAINTED. This will allow the testing of taint propagation,
without having to worry about sanitizers.

Example:
TestingObjectTags.getTaintedInt1(I)I*MAYBE_TAINTED -> explicitly
declared as MAYBE_TAINTED
TestingObjectTags.getTaintedInt1(I)I -> Assumed to be tainted